+++ tests/_spherical/coodinates.R
@@ -2,7 +2,7 @@ library(testthat)
 library(icosa)
 library(restools)
 
-if(getOS()=="windows") setwd("D:/Dropbox/WorkSpace/2014-09-08_icosa/icosa")
+setwd("/mnt/sky/Dropbox/WorkSpace/2014-09-08_icosa/icosa")
 
 set.seed(0)
 xy <- rpsphere(30, output="polar")