+++ tests/generic/bracket_subset/test_bracket.R
@@ -1,0 +1,5 @@
+# thing
+#subscript
+
+expect_silent(subs <- thing[subscript])
+