+++ tests/generic/resample/test_resample.R
@@ -1,0 +1,6 @@
+expect_silent(re <- resample(r, gr))
+expect_equal(sum(re), sol)
+
+
+## fl<-facelayer(gr)
+## fl[] <-re